safeWget cashing

+++ bash-utils/bash-utils.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# QUICK EDIT: FILE="/usr/local/bin/kira-utils.sh" && rm $FILE && nano $FILE && chmod 555 $FILE
+# QUICK EDIT: FILE="/usr/local/bin/bash-utils.sh" && rm $FILE && nano $FILE && chmod 555 $FILE
 # NOTE: For this script to work properly the KIRA_GLOBS_DIR env variable should be set to "/var/kiraglob" or equivalent & the directory should exist
 REGEX_DNS="^(([a-zA-Z](-?[a-zA-Z0-9])*)\.)+[a-zA-Z]{2,}$"
 REGEX_IP="^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$"
@@ -11,58 +11,73 @@ REGEX_PUBLIC_IP='^([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])(?<!172\.(16|1
 REGEX_KIRA="^(kira)[a-zA-Z0-9]{39}$"
 REGEX_VERSION="^(v?)([0-9]+)\.([0-9]+)\.([0-9]+)(-?)([a-zA-Z]+)?(\.?([0-9]+)?)$"
 
-function utilsVersion() {
-    echo "v0.1.1.2"
+
+function bashUtilsVersion() {
+    bashUtilsSetup "version" 2> /dev/null || bash-utils bashUtilsSetup "version"
 }
 
 # this is default installation script for utils
-# ./utils.sh utilsSetup ./utils.sh "/var/kiraglob"
-function utilsSetup() {
-    local UTILS_SOURCE="$1"
-    local GLOBS_DIR="$2"
-    local UTILS_DESTINATION="/usr/local/bin/kira-utils.sh"
-
-    if [ -z "$GLOBS_DIR" ] ; then
-        [ -z "$KIRA_GLOBS_DIR" ] && KIRA_GLOBS_DIR="/var/kiraglob"
+# ./bash-utils.sh bashUtilsSetup "/var/kiraglob"
+function bashUtilsSetup() {
+    local BASH_UTILS_VERSION="v0.1.2.3"
+    if [ "$1" == "version" ] ; then
+        echo "$BASH_UTILS_VERSION"
+        return 0
     else
-        KIRA_GLOBS_DIR=$GLOBS_DIR
-    fi
+        local GLOBS_DIR="$1"
 
-    if [ ! -f $UTILS_SOURCE ] ; then
-        echoErr "ERROR: utils source was NOT found"
-        return 1
-    else
-        mkdir -p "/usr/local/bin"
-        cp -fv "$UTILS_SOURCE" "$UTILS_DESTINATION"
-        cp -fv "$UTILS_SOURCE" "/usr/local/bin/utils"
-        chmod -v 555 $UTILS_DESTINATION "/usr/local/bin/utils"
-        
-        local SUDOUSER="${SUDO_USER}" && [ "$SUDOUSER" == "root" ] && SUDOUSER=""
-        local USERNAME="${USER}" && [ "$USERNAME" == "root" ] && USERNAME=""
-        local LOGNAME=$(logname 2> /dev/null echo "") && [ "$LOGNAME" == "root" ] && LOGNAME=""
+        local UTILS_SOURCE=$(realpath "$0")
+        local VERSION=$($UTILS_SOURCE bashUtilsVersion || echo '')
+        local UTILS_DESTINATION="/usr/local/bin/bash-utils.sh"
 
-        local TARGET="/$LOGNAME/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET="/$USERNAME/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET="/$SUDOUSER/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET="/root/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET=~/.bashrc && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET=~/.zshrc && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-        TARGET=~/.profile && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
-
-        mkdir -p $KIRA_GLOBS_DIR
-
-        utils setGlobEnv KIRA_GLOBS_DIR "$KIRA_GLOBS_DIR"
-        utils setGlobEnv KIRA_TOOLS_SRC "$UTILS_DESTINATION"
-
-        local AUTOLOAD_SET=$(utils getLastLineByPrefix "source $UTILS_DESTINATION" /etc/profile 2> /dev/null || echo "-1")
-
-        if [[ $AUTOLOAD_SET -lt 0 ]] ; then
-            echo "source $UTILS_DESTINATION || echo \"ERROR: Failed to load kira utils from '$UTILS_DESTINATION'\"" >> /etc/profile
+        if [ -z "$GLOBS_DIR" ] ; then
+            [ -z "$KIRA_GLOBS_DIR" ] && KIRA_GLOBS_DIR="/var/kiraglob"
+        else
+            KIRA_GLOBS_DIR=$GLOBS_DIR
         fi
 
-        utils loadGlobEnvs
+        if [ "$VERSION" != "$BASH_UTILS_VERSION" ] ; then
+            bash-utils echoErr "ERROR: Self check version mismatch, expected '$BASH_UTILS_VERSION', but got '$VERSION'"
+            return 1
+        elif [ "$UTILS_SOURCE" == "$UTILS_DESTINATION" ] ; then
+            bash-utils echoErr "ERROR: Installation source script and destination can't be the same"
+            return 1
+        elif [ ! -f $UTILS_SOURCE ] ; then
+            bash-utils echoErr "ERROR: utils source was NOT found"
+            return 1
+        else
+            mkdir -p "/usr/local/bin"
+            cp -fv "$UTILS_SOURCE" "$UTILS_DESTINATION"
+            cp -fv "$UTILS_SOURCE" "/usr/local/bin/bash-utils"
+            chmod -v 555 $UTILS_DESTINATION "/usr/local/bin/bash-utils"
 
-        echoInfo "INFO: SUCCESS!, Installed kira bash-utils $(utilsVersion)"
+            local SUDOUSER="${SUDO_USER}" && [ "$SUDOUSER" == "root" ] && SUDOUSER=""
+            local USERNAME="${USER}" && [ "$USERNAME" == "root" ] && USERNAME=""
+            local LOGNAME=$(logname 2> /dev/null echo "") && [ "$LOGNAME" == "root" ] && LOGNAME=""
+
+            local TARGET="/$LOGNAME/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET="/$USERNAME/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET="/$SUDOUSER/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET="/root/.bashrc" && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET=~/.bashrc && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET=~/.zshrc && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+            TARGET=~/.profile && [ -f $TARGET ] && chmod 777 $TARGET && echoInfo "INFO: /etc/profile executable target set to $TARGET"
+
+            mkdir -p $KIRA_GLOBS_DIR
+
+            bash-utils setGlobEnv KIRA_GLOBS_DIR "$KIRA_GLOBS_DIR"
+            bash-utils setGlobEnv KIRA_TOOLS_SRC "$UTILS_DESTINATION"
+
+            local AUTOLOAD_SET=$(bash-utils getLastLineByPrefix "source $UTILS_DESTINATION" /etc/profile 2> /dev/null || echo "-1")
+
+            if [[ $AUTOLOAD_SET -lt 0 ]] ; then
+                echo "source $UTILS_DESTINATION || echo \"ERROR: Failed to load kira bash-utils from '$UTILS_DESTINATION'\"" >> /etc/profile
+            fi
+
+            bash-utils loadGlobEnvs
+
+            echoInfo "INFO: SUCCESS!, Installed kira bash-utils $(bashUtilsVersion)"
+        fi
     fi
 }
 
@@ -233,13 +248,25 @@ function safeWget() {
     local OUT_PATH=$1
     local FILE_URL=$2
     local EXPECTED_HASH=$3
-    rm -fv $OUT_PATH
+    local OUT_NAME=$(basename $OUT_PATH)
+    local TMP_DIR=/tmp/downloads
+    local TMP_PATH="$TMP_DIR/$OUT_NAME"
 
-    wget "$FILE_URL" -O $OUT_PATH
-    FILE_HASH=$(sha256 $OUT_PATH)
+    mkdir -p $TMP_DIR
+    local FILE_HASH=$(sha256 $TMP_PATH)
+    
+    if [ "$FILE_HASH" == "$EXPECTED_HASH" ] && [ ! -z "$EXPECTED_HASH" ]; then
+        echoInfo "INFO: No need to download, file with the hash '$FILE_HASH' was already found in the '$TMP_DIR' directory"
+        [ "$TMP_PATH" != "$OUT_PATH" ] && cp -fv $TMP_PATH $OUT_PATH
+    else
+        rm -fv $OUT_PATH
+        wget "$FILE_URL" -O $TMP_PATH
+        [ "$TMP_PATH" != "$OUT_PATH" ] && cp -fv $TMP_PATH $OUT_PATH
+        FILE_HASH=$(sha256 $OUT_PATH)
+    fi
 
     if ($(isFileEmpty $OUT_PATH)) ; then
-        echoErr "ERROR: Failed download from '$FILE_URL'"
+        echoErr "ERROR: Failed download from '$FILE_URL', file is exmpty or was NOT found!"
         return 1
     elif [ "$FILE_HASH" != "$EXPECTED_HASH" ]; then
         rm -fv $OUT_PATH || echoErr "ERROR: Failed to delete '$OUT_PATH'"
